+++ v6502/mem.h
@@ -104,7 +104,7 @@ typedef struct {
 	/** @brief Start address of mapped range */
 	uint16_t start;
 	/** @brief Byte-length of mapped range */
-	uint16_t size;
+	size_t size;
     /** @brief Memory access callback for reading bytes within this memory range */
     v6502_readFunction *read;
 	/** @brief Memory access callback for writing bytes within this memory range */
@@ -119,7 +119,7 @@ typedef struct _v6502_memory {
 	/** @brief Memory accessible as a byte-array */
 	uint8_t *bytes;
 	/** @brief Byte-length of memory object */
-	uint16_t size;
+	size_t size;
 	/** @brief Fault Callback Function */
 	void(*fault_callback)(void *context, const char *reason);
 	/** @brief Fault Callback Context */
@@ -142,8 +142,8 @@ typedef struct _v6502_memory {
 
 /** @defgroup mem_lifecycle Memory Lifecycle Functions */
 /**@{*/
-/** @brief Create v6502_memory */
-v6502_memory *v6502_createMemory(uint16_t size);
+/** @brief Create v6502_memory, type is a size_t so that you can alloc an entire 64k with 0x1,0000 */
+v6502_memory *v6502_createMemory(size_t size);
 /** @brief Destroy v6502_memory */
 void v6502_destroyMemory(v6502_memory *memory);
 /** @brief Load a binary blob of expansion ROM into a given v6502_memory */
@@ -154,7 +154,7 @@ void v6502_loadExpansionRomIntoMemory(v6502_memory *memory, uint8_t *rom, uint16
 /**@{*/
 /** @brief Map an address in v6502_memory */
 /** This works by registering an v6502_memoryAccessor as the handler for that range of v6502_memory. Anytime an access is made to that range of memory, the v6502_memoryAccessor is called instead, and is expected to return a byte ready for access. When this function is called, it is also assumed that an access is actually going to happen, which means it is safe to use calls to your callback as trap signals. */
-int v6502_map(v6502_memory *memory, uint16_t start, uint16_t size, v6502_readFunction *read, v6502_writeFunction *write, void *context);
+int v6502_map(v6502_memory *memory, uint16_t start, size_t size, v6502_readFunction *read, v6502_writeFunction *write, void *context);
 /** @brief Read a byte from v6502_memory */
 /** All accesses made by the v6502_cpu should travel through these functions, so that they respect any hardware memory mapping.
 	The trap argument should always be YES when accessed by the CPU, and always NO when accessed by any virtual hardware outside the CPU, or any VM construct (such as logging/introspection mechanisms.)
